# doc/api.yaml
openapi: 3.0.3
info:
  title: Sign2MINT-API
  description: Dokumentation der API für die Sign2MINT-Homepage
  version: 1.0.0
servers:
  - url: '/api/'
...
paths:
...
  /page/{pageName}:
    get:
      summary: "Liefert statische HTML-Seiten"
      operationId: "getPage"
      parameters:
        - name: "pageName"
          in: "path"
          required: true
          schema:
            type: "string"
            example: "about"
      responses:
        "200":
          description: "Das HTML der statischen Seite"
          content:
            application/json:
              schema:
                type: string
                format: "html"
...